rename flare package to proto
